requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-appconf==1.0.1
django-debug-toolbar==1.8
django-extensions==1.7.9
django-image-cropping==1.0.4
easy-thumbnails==2.4.1
graphviz==0.7.1
httplib2==0.10.3
jsonpickle==0.9.4
mongoengine==0.13.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pydot==1.2.3
pydotplus==2.0.2
pymongo==3.4.0
pyparsing==2.2.0
python-openid==2.2.5
python-slugify==1.2.4
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.20
uritemplate==3.0.0
Werkzeug==0.12.2
